Refactor dist op
