+++ Makefile
@@ -22,6 +22,5 @@ lint: cpplint mdlint
 	python3 setup.py lint
 
 postinstall:
-	cd msamp/operators/dist_op && pip install -v . && cd -
-	cd msamp/operators/fp8_op && bash build.sh && cd -
+	cd msamp/operators/dist_op && bash build.sh && cd -
 	cd msamp/optim && pip install -v . && cd -
