+++ msamp/operators/dist_op/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-"""Exposes the interface of MS-AMP dist_op module."""
+"""Exposes the interface of MS-AMP fp8_op module."""
 
 from msamp.operators.dist_op.dist_op import DistOp
