--- msamp/operators/fp8_op/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-"""Exposes the interface of MS-AMP fp8_op module."""
-
-from msamp.operators.fp8_op.fp8_op import FP8Op
-
-__all__ = ['FP8Op']
